* origin/develop: (169 commits)
  Updating nuspec.
  Preparing 4.2.1 release. Bug fix related to ProtoBuf Serialization.
  Random permutations are not part of the project anymore.
  Tabs vs spaces
  Missed some packages.
  Obviously we don't need any orig files here
  Apparently build files got unsynchronized
  Updating readme information
  Updating Release notes
  Enumerable.Empty was not a good idea. Fixing a bug.
  Merging issues detected
  Updating readme information
  Updating Release notes
  Preparing the 4.2.0 release.
  Upgrading Moq to 4.8.x
  Since we don't have DI container in the class libraries, we can remove the good old SoundFingerprinting.Infrastructure. It was a long an productive life we lived together, but it is time to go on different paths.
  Formatting on Fingerprint instance.
  We need to allow unsafe code as we are running some tests with unsafe as well.
  Let's run the FFT on stackallock.
  Bumping to v4.2.0-alpha
  ...

+++ src/SoundFingerprinting.Tests/Unit/DAO/Data/SubFingerprintDataTest.cs
@@ -11,9 +11,9 @@
         [Test]
         public void ShouldIdentifyAsEqual()
         {
-            var dto0 = new SubFingerprintData(new long[0], 0, 0d, new ModelReference<int>(1), new ModelReference<int>(0));
+            var dto0 = new SubFingerprintData(new int[0], 0, 0, new ModelReference<int>(1), new ModelReference<int>(0));
 
-            var dto1 = new SubFingerprintData(new long[0], 0, 0d, new ModelReference<int>(1), new ModelReference<int>(0));
+            var dto1 = new SubFingerprintData(new int[0], 0, 0, new ModelReference<int>(1), new ModelReference<int>(0));
 
             Assert.AreEqual(dto0, dto1);
         }
@@ -21,7 +21,7 @@
         [Test]
         public void ShouldNotBeEqualToNull()
         {
-            var dto = new SubFingerprintData(new long[0], 0, 0d, new ModelReference<int>(1), new ModelReference<int>(0));
+            var dto = new SubFingerprintData(new int[0], 0, 0, new ModelReference<int>(1), new ModelReference<int>(0));
 
             Assert.IsFalse(dto.Equals(null));
         }
